\name{MBCluster.Seq.Internal}
\alias{Hybrid.Tree.Microarray}
\alias{MBCluster.Seq.Internal}
\alias{MI.1}
\alias{MI.2}
\alias{MI.Cluster.Annotation}
\alias{MI.score
}\alias{MI.score.one}
\alias{NMI.score}
\alias{cl.mb}
\alias{cl.mb.est.MC}
\alias{cl.mb.est.P}
\alias{cl.mb.est.mc
}\alias{cl.nb.est.c}
\alias{cl.nb.est.m}
\alias{cl.nb.est.mc}
\alias{cl.ps.est.mc}
\alias{dst.KL}
\alias{dst.Ward
}\alias{dst.euclidean.pair}
\alias{dst.maximum.pair}
\alias{dst.pairs}
\alias{dst.pearson.pair
}\alias{dst2center.pairs}
\alias{est.nb.mu.mle.one}
\alias{est.nb.v.QL}
\alias{est.nb.v.QL.one
}\alias{in.tree}
\alias{leaf.color}
\alias{lglk.cluster}
\alias{lglk.cluster.one}
\alias{lglk.nb}
\alias{lglk.nb.1c
}\alias{lglk.nb.c}
\alias{lglk.ps}
\alias{lglk.ps.1c}
\alias{lglk.ps.c}
\alias{loc.node}
\alias{maxRow}
\alias{meanCol
}\alias{meanRow}
\alias{minRow}
\alias{move.two}
\alias{plotbr}
\alias{sortNode}
\alias{sumCol}
\alias{sumRow}
\alias{tree.K}

\title{
 Internal function for MBCluster.Seq package 
}
\description{
 Internal functions for MBCluster.Seq package 
}
